Move args.rs into the sdk folder
