+++ shared/src/sdk/mod.rs
@@ -1,6 +1,7 @@
 //! Namada's SDK API
 pub mod rpc;
 
+pub mod args;
 pub mod masp;
 pub mod signing;
 #[allow(clippy::result_large_err)]
