+++ shared/src/sdk/signing.rs
@@ -26,7 +26,6 @@ use zeroize::Zeroizing;
 
 use crate::ibc::applications::transfer::msgs::transfer::MsgTransfer;
 use crate::ibc_proto::google::protobuf::Any;
-use crate::ledger::args;
 use crate::ledger::parameters::storage as parameter_storage;
 pub use crate::ledger::wallet::store::AddressVpType;
 use crate::ledger::wallet::{Wallet, WalletUtils};
@@ -34,7 +33,6 @@ use crate::proto::{MaspBuilder, Section, Tx};
 use crate::sdk::masp::{
     make_asset_type, ShieldedContext, ShieldedTransfer, ShieldedUtils,
 };
-use crate::sdk::rpc;
 use crate::sdk::rpc::{
     format_denominated_amount, query_wasm_code_hash, validate_amount,
 };
@@ -44,6 +42,7 @@ use crate::sdk::tx::{
     TX_UNBOND_WASM, TX_UPDATE_ACCOUNT_WASM, TX_VOTE_PROPOSAL, TX_WITHDRAW_WASM,
     VP_USER_WASM,
 };
+use crate::sdk::{args, rpc};
 use crate::types::error::{EncodingError, Error, TxError};
 use crate::types::key::*;
 use crate::types::masp::{ExtendedViewingKey, PaymentAddress};
